@@ -1,21 +1,39 @@
 {
-  "name": "stackblitz-nativescript-typescript",
-  "main": "app/app.ts",
-  "version": "1.0.0",
+  "name": "vite-react-typescript-starter",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint .",
+    "preview": "vite preview"
+  },
   "dependencies": {
-    "@nativescript/core": "~8.8.0",
-    "@nativescript/local-notifications": "^6.1.1",
-    "@nativescript/theme": "~3.0.2",
-    "date-fns": "^2.30.0"
+    "@dnd-kit/core": "^6.1.0",
+    "@dnd-kit/sortable": "^8.0.0",
+    "@dnd-kit/utilities": "^3.2.2",
+    "clsx": "^2.1.0",
+    "date-fns": "^3.3.1",
+    "lucide-react": "^0.344.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "zustand": "^4.5.2"
   },
   "devDependencies": {
-    "@nativescript/preview-cli": "^1.0.14",
-    "@nativescript/stackblitz": "0.0.8",
-    "@nativescript/tailwind": "^2.1.0",
-    "@nativescript/types-minimal": "~8.8.0",
-    "@nativescript/webpack": "~5.0.0",
-    "tailwindcss": "~3.4.0",
-    "typescript": "~5.4.0"
+    "@eslint/js": "^9.9.1",
+    "@types/react": "^18.3.5",
+    "@types/react-dom": "^18.3.0",
+    "@vitejs/plugin-react": "^4.3.1",
+    "autoprefixer": "^10.4.18",
+    "eslint": "^9.9.1",
+    "eslint-plugin-react-hooks": "^5.1.0-rc.0",
+    "eslint-plugin-react-refresh": "^0.4.11",
+    "globals": "^15.9.0",
+    "postcss": "^8.4.35",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.5.3",
+    "typescript-eslint": "^8.3.0",
+    "vite": "^5.4.2"
   }
 }
